Add bullshit device tree stubs.
LG's 4.2 aboot searches for

/soc/qcom,mdss_dsi_lgit_video
/soc/qcom,mdss_dsi_jdi_video
/soc/qcom,mdss_dsi_jdi_cmd
/soc/qcom,mdss_dsi_lgit_cmd

before booting the kernel; and it will explicitly fail to boot if they aren't
present, as defined somewhere in their aboot source. This will appease the
4.2 bootloader and allow us to boot a kernel built on 4.4 sources.

The 4.4 aboot will also append:
"mdss_mdp.panel=1:dsi:0:qcom,mdss_dsi_g2_lgd_cmd" to kernel's command line, so
adding the above to the BOARD_KERNEL_CMDLINE in the device folder is necessary
to get the panel to work. (Thanks to xda user garyd9 for pulling /proc/cmdline)

A big thank you to Dan Rosenberg (djrbliss) for loki (the exploit that allows
for custom boot images)